zeit/requirements.txt
starlette
uvicorn
python-multipart
aiofiles
aiohttp
scikit-image
keras
matplotlib
numpy
tensorflow
h5py
